move crawler into its own module & only spawn necessary tokio tasks
